Linked inventory to account system
