+++ inventory/templates/inventory/membership_take_home.html
@@ -1,4 +1,4 @@
-{% extends 'core/base.html' %}
+{% extends 'core/user_accounts/account_base.html' %}
 {% load static %}
 
 {% block title %}
@@ -43,7 +43,7 @@
                 </div>
                 <div class="col-12">
                     <input type="submit" class="btn btn-warning" form="main_form" value="Yes, I shall take it home">
-                    <a class="btn btn-secondary" href="{% url "inventory:member_items" %}">
+                    <a class="btn btn-secondary" href="{% url "account:inventory:member_items" %}">
                         No, leave it at the Knights
                     </a>
                 </div>
